Share `hab-sup` CLAP config with `hab` to properly handle aliases
This fixes https://github.com/habitat-sh/habitat/issues/3130
