+++ components/hab/src/lib.rs
@@ -21,6 +21,7 @@ extern crate habitat_common as common;
 extern crate habitat_depot_client as depot_client;
 extern crate habitat_api_client as api_client;
 extern crate habitat_http_client as http_client;
+extern crate habitat_sup as sup;
 extern crate handlebars;
 
 extern crate ansi_term;
